packages: source list and version info
chatting with @tianon, sources could be more than a single artifact.
Also, the RPM "release" (and "epoch" for that matter) don't cleanly map
except by having it be a "version" string. Which is really only
comparable within a specific distribution or similar, because while that
version may loosely map to an upstream version, that distro's build may
have patches or build parameters.

Since this information is largely for audit and assurance, it is for
comparison within the channels it was received from. And that channel
data ought to be provided as clearly as possible.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>